Deprecate and remove the Storage class (#89)
* Refactored out the Storage class

- This moves to using the underlying StorageClient classes directly.
- functions like batchRead/Write have been made helpers, with the intent
to migrate them into StorageClient as-needed.
- Added a stub function for each verifiedRead/Write, with a downstream
PR to implement the new signing/verification code (as well as a PR to
move some of that code's dependencies into common- repos.)
- Deletes: InMemoryStorage, Storage (class)
- FileSystemStorage remains due to the init code not having a place to
migrate yet. Once that has a home in either a constructor for
StorageClient or init function for exchange steps.
- Fixes potential race conditions in the batch functions by making the map immutable and enforcing uniqueness in the output label values.

unit tests refactored to cover the new code.