Adds a VerifiedStorageClient wrapper for StorageClient (#100)
This is to ensure that all reads and writes generate, and validate
against, signatures. VerifiedStorageClient (and the VerifiedBlob member
class) are storage agnostics, expecting an initialized StorageClient as
an input. It adds tracking for 2 certificates (one for reading, one for
writing) and a private key used in the validation.

- This will depend on extension functions in consent-signaling-client
- This is also sort of a rollback for my refactoring out Storage.
- Adding the actual consent-signaling dependencies is in a downstream PR
to avoid being blocked on commit. As a result
VerifiedBlob.verifiedRead() and VerifiedStorageClient.createBlobI() are
still stubs.