Adds per-task storage selection to ExchangeTaskExecutor
- Still working on updating tests
- Depends on downstream SecretMap (or similar) implmentation. Currently
running with a normal map.
- The breakdown is as follows:

1. ExchangeTaskExecutor is the main runner
2. ExchangeTaskMapper is the only part of the code that is aware of step
types and requirements. ExchangeTaskExecutor just asks it for the type
of task and special rules that go with that task for a particular
workflow step.
3. CertificateMapper is the only part of the code that interacts with the
certificate service API. ExchangeTaskExecutor just asks it for
certificates for the appropriate exchange.
4. StorageSelector is the only part of the code that knows how to turn the
serialized protos from the (not-yet) secret map to StorageClient
objects. ExchangeTaskExecutor is passed in an instance and just asks it
for VerifiedStorageClients.