change from using ExchangeKey to a more readable format
Since we'll need to take in inputs from outside the exchange (InputTask
targets), we'll need some way to supply those per-exchange inputs
without knowing the current ExchangeId. Using dates is the simplest way
and has the added benefit of making us less reliant on the v2alpha API
spec.

This changes all usages where an ExchangeKey would be expected to use
RecurringExchangeId (an invariant provided in the workflow) and date
(retrieved from ExchangeStep).