Key task logs by new TaskLog CoroutineContext element.
Previous, the coroutine name was used as the key. This was a bit of an
abuse of this because it means that we can't rename coroutines (for
debugging) within an ExchangeTask.

This also stores the logs itself in the CoroutineContext, so we do not
need to keep around a ConcurrentHashMap.