Key task logs by new TaskLog CoroutineContext element. (#219)
Previous, the coroutine name was used as the key. This was a bit of an abuse of CoroutineName because it means that we can't rename coroutines (for eg debugging) within an ExchangeTask.

This also stores the logs itself in the CoroutineContext, so we do not need to keep around a ConcurrentHashMap.