Allow missing queries in decrypt.
In CreateQueries when queries are being equalized across shards, some queries will be dropped if the number of queries assigned to a shard exceeds the permissible amount. When decrypting query results, the dropped queries will be absent, meaning that we can't do a strict join of queries back to the original join keys.