Use breakFusion() judiciously to reduce costs.
This function was initially introduced to encourage the job runner to scale the number of workers between stages, by materializing the output of one stage so that it could be processed by a different number of workers. However, materializing intermediate results and shuffling the data to different workers incurs a massive cost (at least, on GCP). It has been observed that this shuffle cost is the single largest line item for GCP-run exchanges, coming in at around 5x the cost of the next largest line item (which is vCPU usage).

In addition, the --dataflow-max-num-workers flag, which was added after breakFusion(), appears to accomplish the same goal of prompting worker scaling (in most cases), but without the associated shuffle cost.

For these reasons, this PR removes breakFusion() from almost all locations. The only locations where it remains are in CopyToSharedStorageBeamTask and CopyFromSharedStorageBeamTask, because it was observed that without breakFusion(), these do not scale workers (also, their total combined shuffle size is <50MiB).

This has been tested multiple times on GCP with an exchange of 1B synthetic EDP events and 800K MP queries. The impact on the MP alone was the following:

* Reduced per exchange shuffle usage from 46TiB to 5TiB.
* Reduced the sum of time in Beam tasks from 131m to 75m.
* Peak worker count was 25% less after removing breakFusion().